+++ src/model/evidence/reads.rs
@@ -13,21 +13,23 @@ use pairhmm;
 
 pub fn prob_snv(
     record: &bam::Record,
-    cigar: &CigarStringView, // TODO: remove this parameter, passing in the record is enough and removes the necessity of copying over the cigar string before
+    cigar: &CigarStringView,
     start: u32,
     variant: &Variant,
     ref_seq: &[u8]
-) -> Result<(LogProb, LogProb), Box<Error>> { // TODO: needs to be able to return None, when variant position is where the CigarString only consumes reference ('D' - Deletion, 'N' - RefSkip), as those reads are not filtered out before but don't have a read base covering the SNV
+) -> Result<Option<(LogProb, LogProb)>, Box<Error>> {
     if let &Variant::SNV(base) = variant {
         if let Some(qpos) = cigar.read_pos(start, false, false)? {
             let read_base = record.seq()[qpos as usize];
             let base_qual = record.qual()[qpos as usize];
             let prob_alt = prob_read_base(read_base, base, base_qual);
             let prob_ref = prob_read_base(read_base, ref_seq[start as usize], base_qual);
-            Ok((prob_ref, prob_alt))
+            Ok( Some( (prob_ref, prob_alt) ) )
         } else {
-            // such a read should be filtered out before
-            panic!("bug: read does not enclose SNV");
+            // a read that spans an SNV might have the respective position deleted (Cigar op 'D')
+            // or reference skipped (Cigar op 'N'), and the library should not choke on those reads
+            // but instead needs to know NOT to add those reads (as observations) further up
+            Ok( None )
         }
     } else {
         panic!("bug: unsupported variant");
@@ -407,4 +409,95 @@ impl<'a> InsertionEmissionParams<'a> {
 
 impl<'a> pairhmm::EmissionParameters for InsertionEmissionParams<'a> {
     default_emission!();
+}
+
+#[cfg(test)]
+mod tests {
+
+    use super::*;
+    use model;
+
+    use std::str;
+    use rust_htslib::bam::record::{Cigar, CigarString};
+
+    #[test]
+    fn test_prob_snv() {
+        let ref_seq: Vec<u8> = b"CCTATACGCGT"[..].to_owned();
+
+        let mut records: Vec<bam::Record> = Vec::new();
+        let mut qname: &[u8];
+        let mut seq: &[u8];
+
+        // Ignore leading HardClip, skip leading SoftClip, reference nucleotide
+        qname = b"HC_SC_M";
+        let cigar = CigarString( vec![Cigar::HardClip(5), Cigar::SoftClip(2), Cigar::Match(6)] );
+        seq  = b"AATATACG";
+        let qual = [20, 20, 30, 30, 30, 40, 30, 30];
+        let mut record1 = bam::Record::new();
+        record1.set(qname, &cigar, seq, &qual);
+        record1.set_pos(2);
+        records.push(record1);
+
+        // Ignore leading HardClip, skip leading Insertion, alternative nucleotide
+        qname = b"HC_Ins_M";
+        let cigar = CigarString( vec![Cigar::HardClip(2), Cigar::Ins(2), Cigar::Match(6)] );
+        seq  = b"TTTATGCG";
+        let qual = [20, 20, 20, 20, 20, 30, 20, 20];
+        let mut record2 = bam::Record::new();
+        record2.set(qname, &cigar, seq, &qual);
+        record2.set_pos(2);
+        records.push(record2);
+
+        // Matches and deletion before position, reference nucleotide
+        qname = b"Eq_Diff_Del_Eq";
+        let cigar = CigarString( vec![Cigar::Equal(2), Cigar::Diff(1), Cigar::Del(2), Cigar::Equal(5)] );
+        seq  = b"CCAACGCG";
+        let qual = [30, 30, 30, 50, 30, 30, 30, 30];
+        let mut record3 = bam::Record::new();
+        record3.set(qname, &cigar, seq, &qual);
+        record3.set_pos(0);
+        records.push(record3);
+
+        // single nucleotide Deletion covering SNV position
+        qname = b"M_Del_M";
+        let cigar = CigarString( vec![Cigar::Match(4), Cigar::Del(1), Cigar::Match(4)] );
+        seq  = b"CTATCGCG";
+        let qual = [10, 30, 30, 30, 30, 30, 30, 30];
+        let mut record4 = bam::Record::new();
+        record4.set(qname, &cigar, seq, &qual);
+        record4.set_pos(1);
+        records.push(record4);
+
+        // three nucleotide RefSkip covering SNV position
+        qname = b"M_RefSkip_M";
+        let cigar = CigarString( vec![Cigar::Equal(1), Cigar::Diff(1), Cigar::Equal(2), Cigar::RefSkip(3), Cigar::Match(4)] );
+        seq  = b"CTTAGCGT";
+        let qual = [10, 30, 30, 30, 30, 30, 30, 30];
+        let mut record5 = bam::Record::new();
+        record5.set(qname, &cigar, seq, &qual);
+        record5.set_pos(0);
+        records.push(record5);
+
+
+        // truth
+        let probs_ref = [0.9999,   0.00033, 0.99999  ];
+        let probs_alt = [0.000033, 0.999,   0.0000033];
+        let eps       = [0.000001, 0.00001, 0.0000001];
+
+        let vpos = 5;
+        let variant = model::Variant::SNV(b'G');
+        for (i, rec) in records.iter().enumerate() {
+            println!("{}", str::from_utf8(rec.qname()).unwrap());
+            if let Ok( Some( (prob_ref, prob_alt) ) ) = prob_snv(rec, &rec.cigar(), vpos, &variant, &ref_seq) {
+                println!("{:?}", rec.cigar());
+                println!("Pr(ref)={} Pr(alt)={}", (*prob_ref).exp(), (*prob_alt).exp() );
+                assert_relative_eq!( (*prob_ref).exp(), probs_ref[i], epsilon = eps[i]);
+                assert_relative_eq!( (*prob_alt).exp(), probs_alt[i], epsilon = eps[i]);
+            } else {
+                // anything that's tested for the reference position not being covered, should
+                // have 10 as the quality value of the first base in seq
+                assert_eq!(rec.qual()[0], 10);
+            }
+        }
+    }
 }
